Prevent SearchBar form from submitting when submit button is clicked
But why do we use a form if we will just prevent it from being
submitted? Simply because we get the free functionality that the
submit button is activated when a user presses enter on the search
term field. We don't need to create functions that would check if
enter is pressed in the input field.