Import 'connect', 'bindActionCreators', and 'fetchWeather'
The goal here is to call the action creator whenever the user
submits the form (see: TODO in search_bar.js onFormSubmit
function). We need to wire up the form's submit so that whenever
a user types something in and then presses enter or clicks the
search button, we fire that action creator and make our actual
API request. So remember, we're working in a container here and
the container needs to be able to call an action creator. It needs
to be able to reach out to redux and work with redux directly.
That means we need to connect our search bar container to redux
using the connect method from the react-redux library. We're also
going to bind the action creator fetch weather as a property to
this container