Handle the payload in the reducer for action type FETCH_WEATHER
We want to collect the result of different cities, so we have to
include in the returned array the current data in the state. We
could have used:

	return state.push[action.payload.data];

to combine the existing and the new data, but this line of code is
opening up a huge can of worms. This is one of the most subtle
traps inside redux. We should only update state by calling
'this.setState', not manipulate it directly. The above line of
code directly mutates the state. Instead of manipulating the state,
we should return a completely new instance of state. Another way
to do that is by the following code:

	return state.concat[action.payload.data];

This line of code does not mutate the state but returns a new
array with the current values of state concatenated with the
action payload data.